Impact of disabling MSA service on Subscription Activation
Proposed this change as during the troubleshooting of a case we had a customer with 700+ devices that were not upgrading from Win 10 Pro to Ent. Discovered the issue was coming from the fact that the MSA was disabled and we could not generate the  MSA ticket for license authentication that would enable the upgrade to take place. Possible call generator for Windows teams supporting this type of Activation